requirements.txt
django~=3.2
django-cors-headers~=3.13
django-environ~=0.9
django-health-check~=3.17
djangorestframework~=3.13
gunicorn~=20.1
nltk~=3.7
whitenoise~=6.2
